rates_mts/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
gunicorn==20.0.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
cssbeautifier==1.14.9
cssselect==1.2.0
Django==4.2.5
djlint==1.32.1
EditorConfig==0.12.3
filelock==3.12.3
greenlet==2.0.2
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
lxml==4.9.3
packaging==23.1
parsel==1.8.1
pathspec==0.11.2
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyOpenSSL==23.2.0
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
scrapyd==1.4.2
scrapyd-client==1.2.3
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tldextract==3.5.0
tomli==2.0.1
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.7.1
uberegg==0.1.1
urllib3==2.0.4
w3lib==2.1.2
zope.interface==6.0
